paho-mqtt-c and paho-mqtt-cpp: add new packages
The Paho MQTT C Client is a fully fledged MQTT client written in ANSI
standard C. There are two C APIs, "Synchronous" and "asynchronous",
for which the API calls start with MQTTClient and MQTTAsync
respectively.

The Paho MQTT C++ Client provides an interface which is intended to
mirror the Java API as closely as possible. It requires the Paho MQTT
C Client library.

Signed-off-by: Guilherme Maciel Ferreira <guilherme.maciel.ferreira@gmail.com>